Create an adjacency matrix from fiber tracking results

The function takes two objects, fiberobj with class 'dwiFiber' containing fiber tracking results and an array or nifti-object containing atlas information. For each combination of regions defined in the atlas the number of fibers connecting these regions is calculated, resulting in a matrix of fiber counts. As default this matrix is standardized and the diagonal elements are set to zero.

AdjacencyMatrix(fiberobj, atlas, labels = NULL,
  method = c("standardize", "counts"), diagelements = FALSE,
  symmetric=TRUE, verbose = FALSE)

Arguments

fiberobj

an object of class 'dwiFiber'

atlas

an object of class 'array' or 'nifti' containing region indices as intensities. The atlas needs to be registered to DWI (subject) space, with array dimension corresponding to fiberobj@ddim

labels

optional labels for the regions. Will be used as dimnames of the resulting matrix.

method

either "standardize" or "counts", determines if fiber counts or a standardized (default) matix is returned.

diagelements

logical, if FALSE the diagonal elements of the standardized matrix are set to zero (default).

symmetric

logical, with ni the number of fibers originating if FALSE standardized values counts(i,j)/ni, if TRUE we get counts(i,j)/sqrt(nj*nj).

verbose

logical, if TRUE report pairwise fiber counts.

Value

A matrix with dimensions equal to the number of regions defined in the atlas and dimnames given by labels or by the region number. The matrix contains fiber counts or values standardized with the number of fibers ni, nj originating/ending from the pair of regions. Depending on symmetric standardization is with 1/sqrt(ni*nj) or with 1/ni.

Author

Joerg Polzehl polzehl@wias-berlin.de

See also