Estimate noise variance for multicoil MR systems

The distribution of image intensity values \(S_i\) divided by the noise standard deviation in \(K\)-space \(\sigma\) in dMRI experiments is assumed to follow a non-central chi-distribution with \(2L\) degrees of freedom and noncentrality parameter \(\eta\), where \(L\) refers to the number of receiver coils in the system and \(\sigma \eta\) is the signal of interest. This is an idealization in the sense that each coil is assumed to have the same contribution at each location. For realistic modeling \(L\) should be a locally smooth function in voxel space that reflects the varying local influence of the receiver coils in the the reconstruction algorithm used.

The functions assume \(L\) to be known and estimate either a local (function awslsigmc) or global ( function awssigmc) \(\sigma\) employing an assumption of local homogeneity for the noncentrality parameter \(\eta\).

Function afsigmc implements estimates from Aja-Fernandez (2009). Function aflsigmc implements the estimate from Aja-Fernandez (2013).

awssigmc(y, steps, mask = NULL, ncoils = 1, vext = c(1, 1), lambda = 20, 
         h0 = 2, verbose = FALSE, sequence = FALSE, hadj = 1, q = 0.25, 
         qni = .8, method=c("VAR","MAD"))
awslsigmc(y, steps, mask = NULL, ncoils = 1, vext = c(1, 1), lambda = 5, minni = 2, 
         hsig = 5, sigma = NULL, family = c("NCchi"), verbose = FALSE, 
         trace=FALSE, u=NULL)
afsigmc(y, level = NULL, mask = NULL,  ncoils = 1,   vext = c( 1, 1),   
        h = 2, verbose = FALSE, hadj = 1, 
        method = c("modevn","modem1chi","bkm2chi","bkm1chi"))
aflsigmc(y, ncoils, level = NULL, mask = NULL, h=2, hadj=1,  vext = c( 1, 1))

Arguments

y
3D array, usually obtained from an object of class dwi as obj@si[,,,i] for some i, i.e. one 3D image from an dMRI experiment.

steps
number of steps in adapive weights smoothing, used to reveal the unerlying mean structure.

mask
restrict computations to voxel in mask, if is.null(mask) all voxel are used. In function afsigmc mask should refer to background for method %in% c("modem1chi","bkm2chi","bkm1chi") and to voxel within the head for method=="modevn".

ncoils
number of coils, or equivalently number of effective degrees of freedom of non-central chi distribution divided by 2.

vext
voxel extentions

lambda
scale parameter in adaptive weights smoothing

h0
initial bandwidth

verbose
if verbose==TRUE density plots and quantiles of local estimates of sigma are provided.

trace
if trace==TRUE intermediate results for each step are returned in component tergs for all voxel in mask.

sequence
if sequence=TRUE a vector of estimates for the noise standard deviation sigma for the individual steps is returned instead of the final value only.

hadj
adjustment factor for bandwidth (chosen by bw.nrd) in mode estimation

q
quantile to be used for interquantile-differences.

qni
quantile of distribution of actual sum of weights \(N_i=\sum_j w_{ij}\) in adaptive smoothing. Only voxel i with \(N_i > q_{qni}(N_.)\) are used for variance estimation. Should be larger than 0.5.

method
in case of function awssigmc the method for variance estimation, either "VAR" (variance) or "MAD" (mean absolute deviation). In function afsigmc see last column in Table 2 in Aja-Fernandez (2009).

level
threshold for background separation. Used if !is.null(level) to redefine mask

h
bandwidth for local avaeraging

minni
Minimum sum of weights for updating values of sigma.

hsig
Bandwidth of the median filter.

sigma
Initial estimate for sigma

family
One of "Gauss" or "NCchi" (default) defining the probability distribution to use.

u
if verbose==TRUE an array of noncentrality paramters for comparisons. Internal use for tests only

Value

a list with components

sigma

either a scalar or a vector of estimated noise standard deviations.

theta

the estimated mean structure

References

K. Tabelow, H.U. Voss, J. Polzehl, Local estimation of the noise level in MRI using structural adaptation, Medical Image Analysis, 20 (2015), pp. 76--86.

Author

J\"org Polzehl polzehl@wias-berlin.de