Class "dwi"

The family of "dwi" classes is used for Diffusion Weighted Imaging (DWI) data and, within the Diffusion Tensor Model (DTI), diffusion tenors and its indices.

Objects from the Class

"dwi" is only a superclass, no instances should be created. However, objects can be created by calls of the form new("dwi", ...). "dtiData", "dtiTensor", and "dtiIndices" can be created from their correspondingly named functions and methods.

Slots

.Data:

Object of class "list", usually empty.

gradient:

Object of class "matrix", matrix of dimension c(3,ngrad) containing gradient directions.

btb:

Object of class "matrix", matrix of dimension c(6,ngrad) obtained from gradient directions.

bvalue:

Object of class "numeric", of length ngrad containing b-values if available.

ngrad:

Object of class "integer", number of gradients (including zero gradients).

s0ind:

Object of class "integer", index of zero gradients within the sequence 1:ngrad.

replind:

Object of class "integer", index (identifier) of unique gradient directions. Used to characterize replications in the gradient design by identical indices. length is ngrad.

ddim:

Object of class "integer", dimension of subcube defined by xind, yind and zind.

ddim0:

Object of class "integer", dimension of original image cubes. Vector of length 3.

xind, yind, zind:

Objects of class "integer", index for subcube definition in x-, y- and z-direction.

voxelext:

Object of class "numeric", voxel extensions in x-, y- and z-direction. Vector of length 3.

orientation:

Object of class "integer", orientation of data according to AFNI convention. Vector of length 3.

rotation:

Object of class "matrix", optional rotation matrix for gradient directions.

level:

Object of class "numeric", minimal valid S0-level. No evaluation will be performed for voxels with S0-values less than level.

source:

Object of class "character", name of the source imgage file or source directory.

call:

Object of class "call", call that created the object.

For class "dtiData":

si:

Object of class "array", Diffusion Weighted Data.

sdcoef:

Object of class "numeric", Parameters of the model for error standard deviation as a function of the mean. First two entries refer to intercept and slope of a linear function, third and fourth value are the endpoints of the interval of linearity. Contains rep(0,4) if not set. If the function

For class "dtiTensor":

D:

Object of class "array", estimated tensors, dimension c(6,ddim). Tensors are stored as upper diagonal matrices.

th0:

Object of class "array", estimated intensities in S0 images, dimension ddim

sigma:

Object of class "array", estimated error variances if method=="linear", zero otherwise.

scorr:

Object of class "numeric", estimated spatial correlations in coordinate directions

bw:

Object of class "numeric", bandwidth for a Gaussian kernel that approximately creates the estimated spatial correlations. Needed for adjustments of critical values in the adaptive smoothing algorithm used in function dti.smooth

mask:

Object of class "array", logical indicating the voxel where the tensor was estimated.

hmax:

Object of class "numeric", maximal bandwidth in case of adaptive smoothing, 1 otherwise.

outlier:

Object of class "numeric", index of voxel where physical constraints are not met, i.e. where the observed values in gradient images Si were larger than the corresponding S0 values. These are probably motion effects or registration errors. Values are replaced by the corresponding (mean) S0 values.

scale:

Numerical value corresponding to the 95% quantile of the maximal eigenvalues of estimated tensors within the mask. Used for scaling in function show3d.dtiTensor

method:

Object of class "character", either "linear" or "nonlinear" or "unknown". Indicates the regression model used for estimating the tensors.

For class "dtiIndices":

fa:

Object of class "array", Fractional anisotropy values (FA)

ga:

Object of class "array", Geodetic anisotropy values (GA)

md:

Object of class "array", Mean diffusivity values (MD)

andir:

Object of class "array", Main directions of anisotropy

bary:

Object of class "array", Shape parameters

method:

Object of class "character" either "linear" or "nonlinear" or "unknown". Indicates the regression model used for estimating the tensors.

For class "dkiTensor":

D:

Object of class "array", estimated tensors, dimension c(6,ddim). Tensors are stored as upper diagonal matrices.

W:

Object of class "array", estimated kurtosis tensors, dimension c(15,ddim).

th0:

Object of class "array", estimated intensities in S0 images, dimension ddim

sigma:

Object of class "array", estimated error variances if method=="linear", zero otherwise.

scorr:

Object of class "numeric", estimated spatial correlations in coordinate directions

bw:

Object of class "numeric", bandwidth for a Gaussian kernel that approximately creates the estimated spatial correlations. Needed for adjustments of critical values in the adaptive smoothing algorithm used in function dti.smooth

mask:

Object of class "array", logical indicating the voxel where the tensor was estimated.

hmax:

Object of class "numeric", maximal bandwidth in case of adaptive smoothing, 1 otherwise.

outlier:

Object of class "numeric", index of voxel where physical constraints are not met, i.e. where the observed values in gradient images Si were larger than the corresponding S0 values. These are probably motion effects or registration errors. Values are replaced by the corresponding (mean) S0 values.

scale:

Numerical value corresponding to the 95% quantile of the maximal eigenvalues of estimated tensors within the mask. Used for scaling in function show3d.dtiTensor

method:

Object of class "character", either "linear" or "nonlinear" or "unknown". Indicates the regression model used for estimating the tensors.

For class "dkiIndices":

fa:

Object of class "array", Fractional anisotropy values (FA)

ga:

Object of class "array", Geodetic anisotropy values (GA)

md:

Object of class "array", Mean diffusivity values (MD)

andir:

Object of class "array", Main directions of anisotropy

bary:

Object of class "array", Shape parameters

k1:

Object of class "array", Kurtosis along DT (Hui et al. 2008)

k2:

Object of class "array", Kurtosis along DT (Hui et al. 2008)

k3:

Object of class "array", Kurtosis along DT (Hui et al. 2008)

mk:

Object of class "array", Mean kurtosis (Hui et al. 2008)

mk2:

Object of class "array", Mean Kurtosis (Tabesh et al. (2011))

kaxial:

Object of class "array", Axial kurtosis (Hui et al. 2008)

kradial:

Object of class "array", Radial kurtosis (Hui et al. 2008)

fak:

Object of class "array", Kurtosis anisotropy (Hui et al. 2008)

method:

Object of class "character" either "linear" or "nonlinear" or "unknown". Indicates the regression model used for estimating the tensors.

For class "dwiQball":

order:

Object of class "integer", maximal order of Spherical Harmonics to use, needs to be even.

forder:

Object of class "integer", maximal order Gaussian-Laguerre functions in SPF basis (for EAP estimation)

zeta:

Object of class "numeric", Scale parameter used in Gaussian-Laguerre functions (for EAP estimation)

lambda:

Object of class "numeric", nonnegative regularization parameter.

sphcoef:

Object of class "array", estimated coefficients for spherical harmonics, dimension c((order+1)*(order+2)/2,ddim).

sigma:

Object of class "array", estimated error variances if method=="linear", zero otherwise.

scorr:

Object of class "numeric", estimated spatial correlations in coordinate directions

bw:

Object of class "numeric", bandwidth for a Gaussian kernel that approximately creates the estimated spatial correlations. Needed for adjustments of critical values in the adaptive smoothing algorithm used in function dti.smooth

mask:

Object of class "array", logical indicating the voxel where the tensor was estimated.

hmax:

Object of class "numeric", maximal bandwidth in case of adaptive smoothing, 1 otherwise.

outlier:

Object of class "numeric", index of voxel where physical constraints are not met, i.e. where the observed values in gradient images Si were larger than the corresponding S0 values. These are probably motion effects or registration errors. Values are replaced by the corresponding (mean) S0 values.

scale:

Numerical value corresponding to the 95% quantile of the maximal eigenvalues of estimated tensors within the mask. Used for scaling in function show3d.dwiQball

what:

Object of class "character", "ODF", "wODF", "aODF" or "ADC". Indicates if the object contains coefficients of the orientation density function (ODF (Descoteaux 2007), wODF (Sapiro(2009) or aODF) or the apparent diffusion coefficient (ADC). Coefficients are computed with respect to spherical harmonics of the specified order.

For class "dwiFiber":

fibers:

Object of class "matrix", Matrix of fibers. The first three columns contain the coordinates of the track points, the last three columns the direction vectors for each of these points.

startind:

Object of class "integer", indices for the first dimension of fibers where coordinates for a new fiber start.

roix:

Object of class "integer", coordinate range of region of interest in x-direction

roiy:

Object of class "integer", coordinate range of region of interest in x-direction

roiz:

Object of class "integer", coordinate range of region of interest in x-direction

method:

Object of class "character", fiber tracking method.

minfa:

Object of class "numeric", minimal fractional anisotropy index

maxangle:

Object of class "numeric", maximal angle between fibres.

For class "dwiMixtensor":

model:

Object of class "character", characterizes the type of the mixed tensor model. Currently the only implemented model is model="homogeneous_prolate".

ev:

Object of class "array", estimated eigenvalues, dimension c(2,ddim)

mix:

Object of class "array", estimated mixture coefficients, dimension c(nmix,ddim). nmix is the number of mixture components specified.

orient:

Object of class "array", estimated tensor orientations, dimension c(2,nmix,ddim)

th0:

Object of class "array", estimated intensities in S0 images, dimension ddim

sigma:

Object of class "array", estimated error variances if method=="linear", zero otherwise.

scorr:

Object of class "numeric", estimated spatial correlations in coordinate directions

bw:

Object of class "numeric", bandwidth for a Gaussian kernel that approximately creates the estimated spatial correlations. Needed for adjustments of critical values in the adaptive smoothing algorithm used in function dti.smooth

mask:

Object of class "array", logical indicating the voxel where the tensor was estimated.

hmax:

Object of class "numeric", maximal bandwidth in case of adaptive smoothing, 1 otherwise.

outlier:

Object of class "numeric", index of voxel where physical constraints are not met, i.e. where the observed values in gradient images Si were larger than the corresponding S0 values. These are probably motion effects or registration errors. Values are replaced by the corresponding (mean) S0 values.

scale:

Numerical value corresponding to the 95% quantile of the maximal eigenvalues of estimated tensors within the mask. Used for scaling in function show3d.dtiTensor

method:

Object of class "character", either "mixtensor" or "Jian". Indicates the regression model used for estimating the tensors.

Methods

Methods only operate on subclasses "dtiData", "dtiTensor", "dtiIndices", "dwiQball" and "dwiFiber".

dti.smooth

Create estimates of diffusion tensors in each voxel using structural adaptive spatial smoothing.

dtiTensor

signature(object = "dtiData"): Create estimates of diffusion tensors in each voxel.

dtiIndices

signature(object = "dtiTensor"): Create estimates of diffusion tensors indices in each voxel.

tracking

signature(object = "dtiTensor") or signature(object = "dtiIndices"): Fiber tracking.

dtiQball

signature(object = "dtiData"): Create estimates of ADC-parameters with respect to a sherical harmonics ortho-normal system.

show3d

Method for Function `show3d' in Package `dti'.

plot

Method for Function `plot' in Package `dti'.

print

Method for Function `print' in Package `dti'.

summary

Method for Function `summary' in Package `dti'.

References

J. Polzehl and K. Tabelow, Beyond the diffusion tensor model: The package dti, Journal of Statistical Software, to appear.

K. Tabelow, H.U. Voss and J. Polzehl, Modeling the orientation distribution function by mixtures of angular central Gaussian distributions, Journal of Neuroscience Methods, to appear.

J. Polzehl and K. Tabelow, Structural adaptive smoothing in diffusion tensor imaging: The R package dti, Journal of Statistical Software, 31 (2009) pp. 1--24.

K. Tabelow, J. Polzehl, V. Spokoiny and H.U. Voss. Diffusion Tensor Imaging: Structural adaptive smoothing, NeuroImage 39(4), 1763-1773 (2008).

Author

Karsten Tabelow tabelow@wias-berlin.de
J\"org Polzehl polzehl@wias-berlin.de

See also