Methods for Function `dtiIndices' in Package `dti'

The method creates estimates of the fractional anisotropy (FA) and relative anisotropy (RA) indices, the main directions of anisotropy and several statistics used for visualization.

# S4 method for dtiTensor
dtiIndices(object, mc.cores = setCores(,reprt=FALSE))

Arguments

object
Object of class "dtiTensor"

mc.cores
Number of cores to use. Defaults to number of threads specified for openMP, see documentation of package awsMethods. Our experience suggests to use 4-6 cores if available.

Value

An object of class "dtiIndices".

Methods

obj = "ANY"

Returns a warning.

obj = "dtiTensor"

Estimate tensor indices like trace, fractional and geodesic anisotropy, main diffusion direction and shape parameters.

References

J. Polzehl, K. Tabelow (2019). Magnetic Resonance Brain Imaging: Modeling and Data Analysis Using R. Springer, Use R! series. Doi:10.1007/978-3-030-29184-6.

J. Polzehl and K. Tabelow, Beyond the diffusion tensor model: The package dti, Journal of Statistical Software, to appear.

K. Tabelow, H.U. Voss and J. Polzehl, Modeling the orientation distribution function by mixtures of angular central Gaussian distributions, Journal of Neuroscience Methods, to appear.

J. Polzehl and K. Tabelow, Structural adaptive smoothing in diffusion tensor imaging: The R package dti, Journal of Statistical Software, 31 (2009) pp. 1--24.

K. Tabelow, J. Polzehl, V. Spokoiny and H.U. Voss. Diffusion Tensor Imaging: Structural adaptive smoothing, NeuroImage 39(4), 1763-1773 (2008).

http://www.wias-berlin.de/projects/matheon_a3/

See also

Examples

  if (FALSE) demo(dti_art)