Smooth DWI data

Adaptive smoothing of DWI data. Smoothing is performed both in space and on the sphere (e.g. between images obtained for different gradient directions) employing a natural geometrical distance ( in SE(3)). Structural adaptation is used in space only. Method dwi.smooth refers to the original POAS approach for single shell data. Method dwi.smooth.ms implements an improved method that is applicable for both single and multi-shell data.

# S4 method for dtiData
dwi.smooth(object, kstar, lambda=20, kappa0=NULL, mask=NULL, ncoils=1,
                                 sigma=NULL, level=NULL, vred=4, verbose=FALSE, dist=1,
                                 model=c("Gapprox","Gapprox2","Chi","Chi2"))
  # S4 method for dtiData
dwi.smooth.ms(object, kstar, lambda=12, kappa0=.5, ncoils=1,
                                    sigma=NULL, ws0=1, level=NULL, mask = NULL, xind=NULL,
                                    yind=NULL, zind=NULL, verbose=FALSE,
                                    usemaxni=TRUE, memrelease = TRUE)

Arguments

object
Object of class "dtiData"

kstar
Number of steps in structural adaptation

lambda
Scale parameter in adaptation

kappa0
determines amount of smoothing on the sphere. Larger values correspond to stronger smoothing on the sphere. If kappa0=NULL a value is that corresponds to a variace reduction with factor vred on the sphere.

ncoils
Number of coils in MR system

sigma
Error standard deviation. Assumed to be known and homogeneous in the current implementation. A reasonable estimate may be defined as the modal value of standard deviations obtained using method getsdofsb.

level
Threshold for image intensities when setting mask.

mask
Binary 3D image defining a mask

vred
Used if kappa0=NULL to specify the variance reduction on the sphere when suggesting a value of kappa0

xind
index for x-coordinate

yind
index for y-coordinate

zind
index for z-coordinate

verbose
If verbose=TRUE additional reports are given.

dist
Distance in SE3. Reasonable values are 1 (default, see Becker et.al. 2012), 2 ( a slight modification of 1: with k6^2 instead of abs(k6)) and 3 (using a 'naive' distance on the sphere)

model
Determines which quantities are smoothed. Possible values are "Chi" for observed values (assumed to be distributed as noncentral Chi with 2*ncoils degrees of freedom), "Chi2" for squares of observed values (assumed to be distributed as noncentral Chi-squared with 2*ncoils degrees of freedom). "Gapprox" and "Gapprox2" use a Gaussian approximation for the noncentral Chi distribution to smooth ovserved and squared values, respectively.

ws0
Factor to downweight information from S0 images, defaults to 1/numer of s0 images.

usemaxni
If "usemaxni==TRUE" a strikter penalization is used.

memrelease
If "memrelease==TRUE" try to release allocated memory whenever possible.

Value

An object of class "dtiData" with smoothed diffusion weighted images.

Methods

signature(object) = "ANY"

Returns a warning.

signature(object) = "dtiData"

Smoothing of DWI data

See also