Methods for Function `dwiQball' in Package `dti'

The method estimates, in each voxel, the coefficients of an expansion of the apparent diffusion cefficient (ADC) with respect to a apherical harmonics orthonormal system from the DWI data contained in an object of class "dtiData".

# S4 method for dtiData
dwiQball(object, what="wODF", order=4, lambda=0, mask=NULL)

Arguments

object
Object of class "dtiData"

what
Determines quantity to estimate, coefficients ot the orientation density function (ODF) (what="ODF", what="wODF", what="aODF") or the apparent diffusion coefficient (ADC) (what="ADC") with respect to spherical harmonics of the up to the specified order.

order
even integer: maximum order of the sperical harmonics expansion

lambda
nonnegative regularization parameter.

mask
optional brain mask

Value

An object of class "dwiQball".

Methods

obj = "ANY"

Returns a warning.

obj = "dtiData"

Estimate, in each voxel, the coefficients of an expansion of the orientation density function (ODF) or the apparent diffusion coefficient (ADC) with respect to a apherical harmonics orthonormal system. Note that the maxima of the ADC have no direct interpretation as fibre orientations.

References

M. Descoteaux, E. Angelino, S. Fitzgibbons and R. Deriche, Regularized, Fast and Robust Analytical Q-Ball Imaging, Magnetic Resonance Methods, 2007, 58, 497-512.

J. Polzehl, K. Tabelow (2019). Magnetic Resonance Brain Imaging: Modeling and Data Analysis Using R. Springer, Use R! series. Doi:10.1007/978-3-030-29184-6.

See also

Examples

  if (FALSE) demo(dti_art)