Methods for Function `getmask' in Package `dti'

Create a mask containing voxel inside the head

# S4 method for dtiData
getmask(object, level = NULL, prop = 0.4, size = 3)

Arguments

object
an object of class "dtiData"

level
S0 intensity value to be used to discriminate between voxel inside and outside the brain. A good value of level may be determined using method sdpar in advance.

prop
proportion of voxel in test area with s0 value larger than level needed to decide for a voxel inside the brain

size
size of a cube defining a test area

Methods

obj = "ANY"

Returns a warning

obj = "dtiData"

Create a mask containing voxel inside the head

obj = "array"

Create a mask containing voxel inside the head

Value

The function returns an object of class dtiData.

Author

Karsten Tabelow tabelow@wias-berlin.de
J\"org Polzehl polzehl@wias-berlin.de

See also