Diffusion Kurtosis Imaging (DKI)

These methods estimate, in each voxel, the diffusion kurtosis tensor (and the diffusion tensor) and some scalar indices.

# S4 method for class 'dtiData'
dkiTensor(object, method=c("CLLS-QP", "CLLS-H", "ULLS", "QL", "NLR"),
                   sigma=NULL, L=1, mask=NULL,
                   mc.cores=setCores(, reprt=FALSE), verbose=FALSE)
  # S4 method for class 'dkiTensor'
dkiIndices(object, mc.cores=setCores(, reprt=FALSE),
                   verbose=FALSE)

Arguments

object

Object of class "dtiData"

method

Method for tensor estimation. May be "CLLS-QP" for a qudratic programm solution for the constrained optimization (requires package quadprog), "CLLS-H" for a heuristic approximation described in Tabesh et al. (2011), or "ULLS" for an unconstrained linear least squares estimation. "QL" and "NLR" correspond to the use of unconstrained quasi-likelihood and nonlinear regression, respectively.

sigma

Scale parameter of intensity distribution (unprocessed). Used with method="QL" in the calculation of the expected intensity values.

L

Effective number of coils, 2*L are the degrees of freedom of the intensity distribution (unprocessed). The default corresponds, e.g., to the case of a SENSE reconstruction. Used with method="QL" in the calculation of the expected intensity values.

mask

argument to specify a precomputed brain mask

mc.cores

Number of cores to use. Defaults to number of threads specified for openMP, see documentation of package awsMethods. Not yet fully implemented for these methods.

verbose

Verbose mode.

Value

An object of class "dkiTensor" or "dkiIndices".

Methods

signature(object = "ANY")

Returns a warning

signature(object = "dtiData")

The method "dkiTensor" estimates the diffusion kurtosis model, i.e., the kurtosis tensor and the diffusion tensor.

signature(object = "dkiTensor")

The method "dkiIndices" estimates some scalar indices from the kurtosis tensor. The method is still experimental, some quantities may be removed in future versions, other might be included.

References

A. Tabesh, J.H. Jensen, B.A. Ardekani, and J.A. Helpern, Estimation of tensors and tensor-derived measures in diffusional kurtosis imaging, Magnetic Resonance in Medicine, 65, 823-836 (2011).

E.S. Hui, M.M. Cheung, L. Qi, and E.X. Wu, Towards better MR characterization of neural tissues using directional diffusion kurtosis analysis, Neuroimage, 42, 122-134 (2008).

J. Polzehl, K. Tabelow (2019). Magnetic Resonance Brain Imaging: Modeling and Data Analysis Using R. Springer, Use R! series. Doi:10.1007/978-3-030-29184-6.

https://www.wias-berlin.de/projects/matheon_a3/

Author

Karsten Tabelow tabelow@wias-berlin.de

See also