Methods for Function `dti.smooth' in Package `dti'

The function provides structural adaptive smoothing for diffusion weighted image data within the context of an diffusion tensor (DTI) model. It implements smoothing of DWI data using a structural assumption of a local (anisotropic) homogeneous diffusion tensor model (in case a "dtiData"-object is provided). It also implements structural adaptive smoothing of a diffusion tensor using a Riemannian metric (in case a "dtiTensor"-object is given), although we strictly recommend to use the first variant due to methodological reasons.

# S4 method for class 'dtiData'
dti.smooth(object, hmax=5, hinit=NULL, lambda=20, tau=10, rho=1, 
         graph=FALSE,slice=NULL, quant=.8, minfa=NULL, hsig=2.5, 
         lseq=NULL, method="nonlinear", rician=TRUE, 
         niter=5,result="Tensor")

Arguments

object

Either an object of class "dtiData" or an object of class "dtiTensor"

hmax

Maximal bandwidth

hinit

Initial bandwidth (default 1)

lambda

Critical parameter (default 20)

tau

Critical parameter for orientation scores (default 10)

rho

Regularization parameter for anisotropic vicinities (default 1)

graph

"logical": Visualize intermediate results (default FALSE)

slice

slice number, determines the slice used in visualization

quant

determines minfa as corresponding quantile of FA if is.null(minfa)

minfa

minimal anisotropy index (FA) to use in visualization

hsig

bandwidth for presmoothing of variance estimates

lseq

sequence of correction factors for lambda

method

Method for tensor estimation. May be "linear", "nonlinear"

rician

"logical": apply a correction for Rician bias. This is still experimental and depends on spatial independence of errors.

niter

Maximum number of iterations for tensor estimates using the nonlinear model.

result

Determines the created object. Alternatives are "Tensor" for create a dtiTensor-object and "dtiData" for a dtiData-object containing a smoothed data cube.

Methods

object = "ANY"

Returns a warning.

object = "dtiData"

We highly recommend to use the method dti.smooth on DWI data directly, i.e. on an object of class "dtiData", due to methodological reasons, see Tabelow et al. (2008). It is usually not necessary to use any other argument than hmax, which defines the maximum bandwidth of the iteration.

If model=="linear" estimates are obtained using a linearization of the tensor model. This was the estimate used in Tabelow et.al. (2008). model=="nonlinear" uses a nonlinear regression model with reparametrization that ensures the tensor to be positive semidefinite, see Koay et.al. (2006). If varmethod=="replicates" the error variance is estimated from replicated gradient directions if possible, otherwise (default) an estimate is obtained from the residual sum of squares. If volseq==TRUE the sum of location weights is fixed to \(1.25^k\) within iteration \(k\) (does not depend on the actual tensor). Otherwise the ellipsoid of positive location weights is determined by a bandwidth \(h_k = 1.25^(k/3)\).

Value

An object of class dtiTensor.

References

J. Polzehl and K. Tabelow, Beyond the diffusion tensor model: The package dti, Journal of Statistical Software, to appear.

K. Tabelow, H.U. Voss and J. Polzehl, Modeling the orientation distribution function by mixtures of angular central Gaussian distributions, Journal of Neuroscience Methods, to appear.

J. Polzehl and K. Tabelow, Structural adaptive smoothing in diffusion tensor imaging: The R package dti, Journal of Statistical Software, 31 (2009) pp. 1–24.

K. Tabelow, J. Polzehl, V. Spokoiny and H.U. Voss. Diffusion Tensor Imaging: Structural adaptive smoothing, NeuroImage 39(4), 1763-1773 (2008).

https://www.wias-berlin.de/projects/matheon_a3/

Author

Karsten Tabelow tabelow@wias-berlin.de
J\"org Polzehl polzehl@wias-berlin.de

See also