Methods for positive definite EAP and ODF estimation in Package dti

Compute a positive definite estimate of the Ensemble Average Propagator (EAP) and Orientation Density Function (ODF) using the approach of Cjeng et. al (2012).

# S4 method for class 'dtiData'
dwiSqrtODF(object,what="sqrtODF",order=4,forder=1,lambda=0,D0=1.4e-3)

Arguments

object

Object of class "dtiData"

what

Character, currently only "sqrtODF" is possible

order

Even integer, Order of spherical harmonics approximation.

forder

Integer, Order of radial approximation.

lambda

Non-negative, Regularization parameter.

D0

Numeric vector, grid of diffusivity parameters, typically about 1e-3.

Methods

signature(object = "ANY")

Returns a warning.

signature(object = "dtiData")

Compute a positive definite estimate of the Ensemble Average Propagator (EAP) and Orientation Density Function (ODF) using the approach of Cjeng et. al (2012).

Author

Karsten Tabelow tabelow@wias-berlin.de
J\"org Polzehl polzehl@wias-berlin.de

References

J. Cheng, T. Jiang and R. Deriche. Nonnegative Definite EAP and ODF Estimation via a Unified Multi-Shell HARDI Reconstruction, MICCAI 2012.

See also