Methods for Function `plot' in Package `dti'

Visualization of objects of class "dtiData", "dtiIndices", "dtiTensor"and class "dwiMixtensor"

# S4 method for class 'dtiData'
plot(x, y, slice=1, gradient=NULL, view= "axial", show=TRUE,
   density=FALSE, xind=NULL, yind=NULL, zind=NULL, mar=par("mar"), mgp=par("mgp"), ...)
# S4 method for class 'dtiTensor'
plot(x, y, slice=1, view="axial", quant=0, minfa=NULL, contrast.enh=1,
    what="fa", qrange=c(.01,.99), xind=NULL, yind=NULL, zind=NULL,
    mar=par("mar"), mgp=par("mgp"), ...)
# S4 method for class 'dwiMixtensor'
plot(x, y, slice=1, view="axial", what="fa", minfa=NULL,
     identify=FALSE, xind=NULL, yind=NULL, zind=NULL, mar=par("mar"), mgp=par("mgp"), ...)
# S4 method for class 'dtiIndices'
plot(x, y, slice=1, view= "axial", method=1, quant=0, minfa=NULL,
   show=TRUE, identify=FALSE, density=FALSE, contrast.enh=1, what="fa",
   xind=NULL, yind=NULL, zind=NULL, mar=par("mar"), mgp=par("mgp"), ...)
# S4 method for class 'dwiFiber'
plot(x, y, ...)
# S4 method for class 'dkiIndices'
plot(x, y, slice=1, what=c("md", "fa", "mk", "mk2",
     "k1", "k2", "k3", "kaxial", "kradial", "fak"), xind=NULL, yind=NULL,
     mar=par("mar"), mgp=par("mgp"), ...)

Arguments

x

Object of class "dtiIndices", "dtiData" or "dtiTensor"

y

Not used

slice

Slice number

view

Choose "sagittal", "coronal", or "axial" view here

gradient

Index of data cube to plot. Defaults to the first S0 image.

method

Method for color coding tensor indices.

quant

If is.null(minfa) specify minfa as corresponding quantile of the fractal anisotropy (FA) index.

minfa

Display only information for voxel with (G)FA>minfa

show

Visualize information in a graphics device (for classes "dtiData" and "dtiIndices" only).

identify

Enable identification of coordinates by mouse actions, logical with default FALSE. Uses function identify. (for classes "dtiIndices" and "dwiMixtensor" only)

density

Show density of S0(Sb)-values (for class "dtiData") or densities of fractal anisotropy (FA) or geodesic anisotropy (GA) ( for class "dtiIndices").

contrast.enh

Enhance image contrast using min(1,x$anindex/contrast.enh instead of the anisotropy index itself. Effective values are within the interval (0,1).

what

In case of class "dtiIndices" what="ga" uses geodesic anisotropy (GA) in contrast to what="fa" for fractional anisotropy (FA). For class "dwiMixtensor" what="fa" for FA and what="order" for the number of mixture components may be chosen.

mar

Graphical parameter for par.

mgp

Graphical parameter for par.

qrange

Cut image intensity to these quantiles to avoid that outliers determine the dynamic range of the image.

xind

If provided restrict display to indices specified in xind for x-direction.

yind

If provided restrict display to indices specified in yind for y-direction.

zind

If provided restrict display to indices specified in zind for z-direction.

...

currently not used

Methods

x = "ANY"

Generic function: see plot.

x = "dwi"

Returns a warning.

x = "dtiData"

gradient can be used to specify a specific data cube associated with the index of a gradient direction. For objects of class "dtiData" images are produces that are scaled by the maximal observed image value. This guarantees that subsequently produced images are on a comparable grey scale. The resulting image of class "adimpro" from package adimpro is returned.

x = "dtiIndices"

Color coded anisotropy maps are produced depending on the specification in method. method==1, method==2, method==4 and method==5 specify three different color schemes for directional FA-maps. method==6 uses colored FA maps based on scheme developed at Uni Muenster (M. Deppe, Germany). method==3 specifies visualization of dtiIndices using color coded shape parameters. If identify==FALSE the resulting image of class "adimpro" from package adimpro, otherwise a matrix with coordinates of identified voxel is returned.

x = "dtiTensor"

The tensor itself, fractional anisotropy (FA), mean diffusivity (MD) and a color coded anisotropy map are provided. NULL is returned.

x = "dwiMixtensor"

Depending of what images of FA (what="fa"), number of mixture components (what="order"), effective order (what="eorder") or maximum eigenvalues (what="ev"). is returned.

x = "dwiFiber"

Creates a density plot of fiber lengths. NULL is returned.

x = "dkiIndices"

Prelimanary function to plot a slice of diffusion kurtosis indices: Mean Kurtosis what="mk" or what="mk2", mean diffusivity what="md", fractional anisotropy what="fa".

Author

Karsten Tabelow tabelow@wias-berlin.de
J\"org Polzehl polzehl@wias-berlin.de

See also

Examples

  if (FALSE) demo(dti_art) # \dontrun{}