Methods for Function `tracking' in Package `dti'

The function provides fiber tracking of "dtiTensor", "dtiIndices", and "dwiMixtensor" objects and methods for fiber manipulations.

# S4 method for dtiTensor
tracking(obj, roix=NULL, roiy=NULL, roiz=NULL, mask=NULL,
                    method="LINEPROP", minfa=0.3, maxangle=30, subsample = 1)
# S4 method for dtiIndices
tracking(obj, roix=NULL, roiy=NULL, roiz=NULL, mask=NULL,
                    method="LINEPROP", minfa=0.3, maxangle=30, subsample = 1)
# S4 method for dwiMixtensor
tracking(obj, roix=NULL, roiy=NULL, roiz=NULL, mask=NULL,
                    method="LINEPROP", minfa=0.3, maxangle=30, subsample = 1,
                    mincompartsize = 0)
# S4 method for dwiFiber
selectFibers(obj, roix=NULL, roiy=NULL, roiz=NULL, mask=NULL,
                    minlength=1)
# S4 method for dwiFiber
reduceFibers(obj, maxdist=1, ends=TRUE)

# S4 method for dwiFiber,dwiFiber
combineFibers(obj, obj2)

# S4 method for dwiFiber,dwiFiber
touchingFibers(obj, obj2, maxdist=1, combine=FALSE)

Arguments

obj
An object of class "dtiTensor", "dtiIndices", or "dwiMixtensor" for tracking() and "dwiFiber" for selectFiber(), conbineFibers().

obj2
An object of class "dwiFiber" for combineFibers().

roix
Indices defining the ROI in x direction. Currently min/max is used to define ROIx

roiy
Indices defining the ROI in y direction. Currently min/max is used to define ROIy

roiz
Indices defining the ROI in z direction. Currently min/max is used to define ROIz

mask
Mask defining seed points for tracking

method
Method for fibre tracking. "LINEPROP" is simple line propagation algorithm which is the default.

minfa
Minimal FA to follow the tracks. default 0.3

maxangle
Maximal angle between fiber in adjacent voxels. default 30 degree.

subsample
Subsampling order of the data to get more dense fibre tracks. Note, that objects become very(!) large.

minlength
Minimal length of fibers to be selected.

maxdist
Maximal supremum distance between fibers in mm

ends
Logical: Use only endpoints of shorter fibers for distance (TRUE) or compute distances using full fiber-length (FALSE). Default (TRUE) removes more fibers and is significantly faster.

mincompartsize
Minimal size of a compartment in dwiMixtensor that will be used in fiber tracking.

combine
If combine=TRUE fibers selected from obj are combined with the fibers from obj2.

Methods

obj = "dtiTensor"

Fiber tracking is performed on the estimated vector field of principal diffusion direction using the method method. Currently only line propagation is implemented. The resulting tracks can be visualized using function show3d

.
obj = "dtiIndices"

Fiber tracking is performed on the estimated vector field of principal diffusion direction using the method method. Currently only line propagation is implemented. The resulting tracks can be visualized using function show3d

.
obj = "dwiMixtensor"

Fiber tracking is performed on the estimated vector fields of diffusion direction in the mixed tensor model using the method method. Currently only line propagation is implemented. The resulting tracks can be visualized using function show3d

.
obj = "dwiFiber"

selectFibers produces a dwiFiber-object containing all fibers that cross the region of interest and exceed a minimum length. reduceFibers eliminates all fibers that are within a maximum supremum distance of maxdist mm of a longer fiber. reduceFibers allows to reduce the size of a dwiFiber-object considerably but is slow !

signature(obj1 = "dwiFiber", obj2 = "dwiFiber")

combineFibers produces a dwiFiber-object containing all fibers that are in one of the supplied objects. touchingFibers takes all fibers from obj that have a minimum distance to a fiber in obj2 of less than maxdist. If combine=TRUE these fibers are combined with the fibers from obj2.

Value

The function returns an object of class dwiFiber.

References

J. Polzehl, K. Tabelow (2019). Magnetic Resonance Brain Imaging: Modeling and Data Analysis Using R. Springer, Use R! series. Doi:10.1007/978-3-030-29184-6.

Author

Karsten Tabelow tabelow@wias-berlin.de, Joerg Polzehl polzehl@wias-berlin.de

See also